fix: resolve production blockers - version endpoint, route redirects
- Add /api/version endpoint for deployment verification (Git SHA, build ID)
- Add non-www to www redirect preserving path and query string
- Fix /terms redirect to canonical /terms-of-service
- Fix /accessibility/accessibility redirect to canonical /accessibility
- Add /api/version to public paths in middleware

These changes address PR #489 comments about missing version endpoint
and route reconciliation requirements.
